+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 <!-- Please update cartridge-cli/VERSION.lua with new release -->
 
+### Added
+
+- Cartridge Stateboard support:
+  * Application template contains stateboard entrypoint script and configuration
+  * Unit file for stateboard `systemd` service is delivered in RPM/DEB
+  * Added `--stateboard` and `--stateboard-only` options for `start` and `stop`
+    commands to start/stop stateboard locally
+
 ### Changed
 
 - Prettified `start` and `stop` logs
@@ -15,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   errors
 - If instance is already stopped, `stop` command doesn't fail, only warning
   message is shown
+- Update `cartridge` to 2.1.1
 
 ## [1.7.0] - 2020-04-10